chore: LightdashVisualization utils
