+++ packages/frontend/src/components/LightdashVisualization/VisualizationBigNumberConfig.tsx
@@ -1,20 +1,11 @@
 import { ChartType, ItemsMap } from '@lightdash/common';
 import { FC, useEffect } from 'react';
 import useBigNumberConfig from '../../hooks/useBigNumberConfig';
-import {
-    VisualizationConfig,
-    VisualizationConfigCommon,
-} from './VisualizationProvider';
+import { VisualizationConfigCommon } from './VisualizationProvider';
 
 export type VisualizationConfigBigNumber = {
     chartType: ChartType.BIG_NUMBER;
     chartConfig: ReturnType<typeof useBigNumberConfig>;
-};
-
-export const isBigNumberVisualizationConfig = (
-    visualizationConfig: VisualizationConfig | undefined,
-): visualizationConfig is VisualizationConfigBigNumber => {
-    return visualizationConfig?.chartType === ChartType.BIG_NUMBER;
 };
 
 type VisualizationBigNumberConfigProps =
