+++ packages/frontend/src/components/LightdashVisualization/VisualizationConfigTable.tsx
@@ -1,20 +1,11 @@
 import { ChartType, DashboardFilters, ItemsMap } from '@lightdash/common';
 import { FC, useEffect } from 'react';
 import useTableConfig from '../../hooks/tableVisualization/useTableConfig';
-import {
-    VisualizationConfig,
-    VisualizationConfigCommon,
-} from './VisualizationProvider';
+import { VisualizationConfigCommon } from './VisualizationProvider';
 
 export type VisualizationConfigTable = {
     chartType: ChartType.TABLE;
     chartConfig: ReturnType<typeof useTableConfig>;
-};
-
-export const isTableVisualizationConfig = (
-    visualizationConfig: VisualizationConfig | undefined,
-): visualizationConfig is VisualizationConfigTable => {
-    return visualizationConfig?.chartType === ChartType.TABLE;
 };
 
 type VisualizationTableConfigProps =
