+++ packages/frontend/src/components/LightdashVisualization/VisualizationCustomConfigProps.tsx
@@ -1,20 +1,11 @@
 import { ChartType } from '@lightdash/common';
 import { FC } from 'react';
-import {
-    VisualizationConfig,
-    VisualizationConfigCommon,
-} from './VisualizationProvider';
+import { VisualizationConfigCommon } from './VisualizationProvider';
 
 // TODO: this is a placeholder for custom visualizations
 export type VisualizationConfigCustom = {
     chartType: ChartType.CUSTOM;
     chartConfig: { validConfig: {} };
-};
-
-export const isCustomVisualizationConfig = (
-    visualizationConfig: VisualizationConfig | undefined,
-): visualizationConfig is VisualizationConfigCustom => {
-    return visualizationConfig?.chartType === ChartType.CUSTOM;
 };
 
 type VisualizationCustomConfigProps =
