+++ packages/frontend/src/hooks/echarts/useEchartsCartesianConfig.ts
@@ -38,7 +38,7 @@ import groupBy from 'lodash/groupBy';
 import toNumber from 'lodash/toNumber';
 import moment from 'moment';
 import { useMemo } from 'react';
-import { isCartesianVisualizationConfig } from '../../components/LightdashVisualization/VisualizationConfigCartesian';
+import { isCartesianVisualizationConfig } from '../../components/LightdashVisualization/utils';
 import { useVisualizationContext } from '../../components/LightdashVisualization/VisualizationProvider';
 import { defaultGrid } from '../../constants';
 import { EMPTY_X_AXIS } from '../cartesianChartConfig/useCartesianChartConfig';
