+++ packages/frontend/src/hooks/echarts/useEchartsPieConfig.ts
@@ -1,7 +1,7 @@
 import { formatItemValue, ResultRow, ResultValue } from '@lightdash/common';
 import { EChartsOption, PieSeriesOption } from 'echarts';
 import { useMemo } from 'react';
-import { isPieVisualizationConfig } from '../../components/LightdashVisualization/VisualizationConfigPie';
+import { isPieVisualizationConfig } from '../../components/LightdashVisualization/utils';
 import { useVisualizationContext } from '../../components/LightdashVisualization/VisualizationProvider';
 
 export type PieSeriesDataPoint = NonNullable<
